+++ tests/test-681real_audio_all_timestamps_zero.rb
@@ -1,0 +1,5 @@
+#!/usr/bin/ruby -w
+
+# T_681real_audio_all_timestamps_zero
+describe "mkvmerge / RealAudio all timestamps zero"
+test_merge "data/rm/all_timestamps_zero.rm"